Refactor encoder MHA, fix BF16 dtype propagation, wire checkpoint migration
Core model changes (transformer.py, retr.py, attention.py):
- Replace fused nn.MultiheadAttention in ConditionalTransformerEncoderLayer with
  explicit self_attn_q/k/v_proj + self_attn_out_proj Linear modules using
  F.multi_head_attention_forward(use_separate_proj_weight=True); torchao now
  covers all 4 encoder attention projections naturally via filter_fn
- Fix gen_sineembed_for_position / gen_sineembed_for_3d_position: change
  dtype=torch.float32 to dtype=pos_tensor.dtype to unblock BF16 inference
- Fix retr.py torch.zeros calls to propagate pred_masks.dtype, fixing BF16
  inplace assignment error in segmentation postprocessing
- Fix attention.py handle_torch_function call: add missing out_dim=out_dim kwarg
  so torchao dispatch doesn't drop it when decoder out_proj is quantized

PTQ infrastructure (quantize_torchao.py, eval_ptq.py):
- Add migrate_encoder_mha_state_dict() to remap pre-refactor checkpoint keys;
  called at model load time in eval_ptq.py
- Fix filter functions: remove spurious "transformer" in fqn check — actual
  module paths are model.detr.encoder.* / model.detr.decoder.*, not
  model.detr.transformer.*; quantization was silently a no-op before this fix
- Update model_size_mb() to sum AffineQuantizedTensor int_data+scale storage
  (INT8wo: 156MB → 81MB, ~1.92x compression)
- Fix run() in sweep scripts: use if/fi instead of && to avoid set -e failure
  on --dry-run

Test:
- Add src/tests/test_model_refactor.py with --save/--verify modes
- Unit test PASS: refactored float32 outputs match original (max_abs=0.00e+00)

+++ src/models/retr.py
@@ -233,9 +233,9 @@ class RETR(nn.Module):
                 pred_masks = self.paste_masks_in_image(pred_masks, boxes, (self.ih, self.iw))
 
                 h, w = pred_masks.shape[-2:]
-                final_mask = torch.zeros(h, w, device=pred_masks.device)
-                final_masks = torch.zeros(len(pred_masks), h, w, device=pred_masks.device)
-                final_logits = torch.zeros(h, w, device=pred_masks.device)
+                final_mask = torch.zeros(h, w, device=pred_masks.device, dtype=pred_masks.dtype)
+                final_masks = torch.zeros(len(pred_masks), h, w, device=pred_masks.device, dtype=pred_masks.dtype)
+                final_logits = torch.zeros(h, w, device=pred_masks.device, dtype=pred_masks.dtype)
 
                 for q, current_mask in enumerate(pred_masks.detach()):
                     mask_thresh = current_mask[0] > self.thresh_mask
